+++ tests/Nexendrie/Utils/BackedEnum.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Utils;
+
+enum BackedEnum: string {
+  case ABC = "abc";
+  case DEF = "def";
+  case GHI = "ghi";
+}
+?>