+++ tests/Nexendrie/Utils/BasicEnum.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Utils;
+
+enum BasicEnum {
+  case ABC;
+  case DEF;
+  case GHI;
+}
+?>